src/redis_deployer/requirements.txt
cloudshell-shell-core>=2.3.0,<2.4.0
cloudshell-automation-api>=7.1.0,<7.2.0
azure==2.0.0rc6
